refactor: Move car and layout components into folders for maintainbility
